% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_ext.R
\name{add_time_ext}
\alias{add_time_ext}
\title{Add time information, often as version, to arbitrary file name}
\usage{
add_time_ext(filename, type = c("sec", "min", "day"))
}
\arguments{
\item{filename, }{a file name string}

\item{type, }{accurate to day, min or second, with second as default}
}
\value{
new file name with time information
}
\description{
Add time information, often as version, to arbitrary file name
}
\details{
add time information to distinguish different versions of files
}
