% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_arithmetic.R
\name{bound_arithmetic}
\alias{bound_arithmetic}
\title{bound of covariance - arithmetic mean of variance X and variance Y}
\usage{
bound_arithmetic(x, y)
}
\arguments{
\item{x, }{variable x}

\item{y, }{variable y}
}
\value{
a numeric vector of length one
}
\description{
arithmetic mean of variance X and variance Y
}
\details{
arithmetic mean of variance X and variance Y 
is defined as (var(x)+var(y)) / 2
}
