% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_geometric.R
\name{bound_geometric}
\alias{bound_geometric}
\title{bound of covariance - geometric mean of variance X and variance Y}
\usage{
bound_geometric(x, y)
}
\arguments{
\item{x, }{variable x}

\item{y, }{variable y}
}
\value{
a numeric vector of length one
}
\description{
geometric mean of variance X and variance Y
}
\details{
geometric mean of variance X and variance Y 
is defined as sqrt(var(x)*var(y))
}
