% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_harmonic.R
\name{bound_harmonic}
\alias{bound_harmonic}
\title{bound of covariance - harmonic mean of variance X and variance Y}
\usage{
bound_harmonic(x, y)
}
\arguments{
\item{x, }{variable x}

\item{y, }{variable y}
}
\value{
a numeric vector of length one
}
\description{
harmonic mean of variance X and variance Y
}
\details{
harmonic mean of variance X and variance Y 
is defined as 2/(1/var(x) + 1/var(y))
}
