% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_quadratic.R
\name{bound_quadratic}
\alias{bound_quadratic}
\title{bound of covariance - quadratic mean of variance X and variance Y}
\usage{
bound_quadratic(x, y)
}
\arguments{
\item{x, }{variable x}

\item{y, }{variable y}
}
\value{
a numeric vector of length one
}
\description{
quadratic mean of variance X and variance Y
}
\details{
quadratic mean of variance X and variance Y 
is defined as sqrt((var(x)^2+var(y)^2)/2)
}
