% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_power.R
\name{cal_power}
\alias{cal_power}
\title{Calculate the power}
\usage{
cal_power(power_df, alpha = 0.95)
}
\arguments{
\item{power_df, }{data.frame, the result of sim_power()}

\item{alpha, }{the power threshold, with 1-0.05 = 0.95 as default}
}
\value{
a data.frame
}
\description{
Power calculation on the basis of the results of power analysis
}
\details{
The empirical power at type I error level alpha=0.05 is calculated as 
the fraction of dependent variable pairs yielding a statistic value 
greater than 95% of the values yielded by the independent variable pairs.

Here, alpha should be set as 0.95 rather than 0.05
}
