% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_R2_SNR.R
\name{cal_R2_SNR}
\alias{cal_R2_SNR}
\title{Calculate R2 with certain y_hat and y}
\usage{
cal_R2_SNR()
}
\arguments{
\item{y, }{noised signal}

\item{y_hat, }{signal}

\item{noise, }{noise}
}
\value{
power, scalar value
}
\description{
calculate R2 with certain y_hat and y
}
\details{
The following relationship is assumed: 

y <- y_hat + noise

That is, noise is independent sample added to signal y_hat.

Any two of the three arguments, y, y_hat, noise should be set.

There are different definitions of $\eqn{R^2}
For more details, see T.O. Kvålseth, Cautionary Note about \eqn{R^2}, The American Statistician. 39 (1985) 279–285. https://doi.org/10.1080/00031305.1985.10479448.
}
