% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_ACE.R
\name{cor_ACE}
\alias{cor_ACE}
\title{Alternating Conditional Expectations - Maximal Correlation}
\usage{
cor_ACE(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
Uses the alternating conditional expectations algorithm to 
find the transformations of y and x that 
maximize the proportion of variation in y explained by x.
}
\details{
See help file for acepack::ace()
}
\references{
L. Breiman, J. H. Friedman, 
Estimating Optimal Transformations for Multiple Regression and Correlation. 
Journal of the American Statistical Association. 80, 580–598 (1985)
}
