% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_Ahat.R
\name{cor_Ahat}
\alias{cor_Ahat}
\title{General Multidimensional Associations, or Generalized R-squared}
\usage{
cor_Ahat(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for matie::ma()
}
\details{
See help file for matie::ma
}
\references{
Murrell, B., Murrell, D. & Murrell, H. 
Discovering General Multidimensional Associations. 
PLoS ONE 11, e0151551 (2016). Also available at: 
http://arxiv.org/abs/1303.1828

Q. Tan, M. Thomassen, M. Burton, K. F. Mose, K. E. Andersen, 
J. Hjelmborg, T. Kruse, Generalized Correlation Coefficient 
for Non-Parametric Analysis of Microarray Time-Course Data. 
Journal of Integrative Bioinformatics. 14 (2017), doi:10.1515/jib-2017-0011.
}
