% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_bcor.R
\name{cor_bcor}
\alias{cor_bcor}
\title{Ball Correlation Statistics}
\usage{
cor_ball(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for ball() in Ball::bcor
}
\details{
See help files for Ball::bcor
}
