% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_bicor.R
\name{cor_bicor}
\alias{cor_bicor}
\title{Biweight Midcorrelation}
\usage{
cor_bicor(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for WGCNA::bicor
}
\details{
See help files for WGCNA::bicor
}
\references{
R. R. Wilcox, Introduction to Robust Estimation and Hypothesis Testing, 
3rd edtion. Academic Press, 2012. Chapter 9. pp.446

Peter Langfelder, Steve Horvath (2012) Fast R Functions for 
Robust Correlations and Hierarchical Clustering. 
Journal of Statistical Software, 46(11), 1-17.
 https://www.jstatsoft.org/v46/i11/
}
