% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_biweight.R
\name{cor_biweight}
\alias{cor_biweight}
\title{Tukey's biweight}
\usage{
cor_tic(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for biwt::biwt.cor()
}
\details{
See help files for biwt::biwt.cor()
}
\references{
J. Hardin, A. Mitani, L. Hicks, B. VanKoten, 
A robust measure of correlation between two genes on a microarray. 
BMC Bioinformatics. 8, 220 (2007).
}
