% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_blomqvist.R
\name{cor_blomqvist}
\alias{cor_blomqvist}
\title{Blomqvist's beta}
\usage{
cor_blomqvist(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for wdm::wdm(x, y = NULL, method = "blomqvist", weights = NULL, remove_missing = TRUE)
}
\details{
See help files for wdm::wdm
}
\references{
N. Blomqvist, On a measure of dependence between two random variables. 
The Annals of Mathematical Statistics, 593–600 (1950).
}
