% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_Ccor.R
\name{cor_Ccor}
\alias{cor_Ccor}
\title{Robust Copula Dependence - Copula correlation}
\usage{
cor_mi(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper of rcd::rcd
}
\details{
See help file for rcd::rcd
}
\references{
A.A. Ding, Y. Li, Copula Correlation: 
An Equitable Dependence Measure and Extension of 
Pearson’s Correlation, (2015). http://arxiv.org/abs/1312.7214
}
