% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_CoS.R
\name{cor_CoS}
\alias{cor_CoS}
\title{Copula Statistic}
\usage{
cor_CoS(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
Copula Statistic
}
\details{
Codes from  
https://github.com/stochasticresearch/copulastatistic
}
\references{
M. B. Hassine, L. Mili, K. Karra, 
A Copula Statistic for Measuring Nonlinear Multivariate Dependence (2016), 
(available at http://arxiv.org/abs/1612.07269)
}
