% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_dcor.R
\name{cor_dcor}
\alias{cor_dcor}
\title{Distance correlation coefficient}
\usage{
cor_dcor(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for energy::dcor
}
\details{
More efficient dcor2d are used

Since dcor2d(x, y) = dcor(x, y)^2, its square root is adpoted.
}
