% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_dcor_rank.R
\name{cor_dcor_rank}
\alias{cor_dcor_rank}
\title{Distance Correlation Coefficient - rank version}
\usage{
cor_dcor(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for energy::dcor
}
\details{
calculate cor_dcor on rank(x) and rank(y)

Fast rank:
\itemize{
  \item data.table::frank
  \item pracma::Rank
  \item DescTools::Rank
  \item Rfast::Rank
}

Rfast::Rank is adpoted here since it is the fastest
}
