% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_dhsic.R
\name{cor_dhsic}
\alias{cor_dhsic}
\title{d-variable Hilbert Schmidt independence criterion - dHSIC}
\usage{
cor_hsic(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for dHSIC::dhsic
}
\details{
See help file for dHSIC::dhsic
}
