% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gaussrank.R
\name{cor_gaussrank}
\alias{cor_gaussrank}
\title{Gauss Rank Correlation Estimator}
\usage{
cor_gaussrank(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for rococo::gauss.cor()
}
\details{
See help file for rococo::gauss.cor
}
\references{
BOUDT K, CORNELISSEN J, CROUX C. 
The Gaussian rank correlation estimator: robustness properties[J/OL]. 
Statistics and Computing, 2012, 22(2): 471-483. DOI:10.1007/s11222-011-9237-0
}
