% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gmic.R
\name{cor_gmic}
\alias{cor_gmic}
\title{The Generalized Mean Information Coefficient}
\usage{
cor_gmic(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for minerva::mine_stat(x, y, measure = 'gmic')
}
\details{
See help file for minerva::mine_stat
}
\references{
D. Albanese, M. Filosi, R. Visintainer, S. Riccadonna, G. Jurman, C. Furlanello. 
minerva and minepy: a C engine for the MINE suite and its R, Python and MATLAB wrappers. 
Bioinformatics (2013) 29(3): 407-408
}
