% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_Hellinger.R
\name{cor_Hellinger}
\alias{cor_Hellinger}
\title{The Hellinger Correlation}
\usage{
cor_ball(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper of HellCor::HellCor
Empirical value of the Hellinger correlation between two continuous random variables.
}
\details{
See help file for HellCor::HellCor
}
\references{
G. Geenens, P. Lafaye de Micheaux, The Hellinger Correlation. 
Journal of the American Statistical Association, 1–15 (2020)
}
