% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_HHG.R
\name{cor_HHG}
\alias{cor_HHG}
\title{Heller-Heller-Gorfine (HHG) Tests of Independence}
\usage{
cor_HHG(x, y, normalized = TRUE)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}

\item{normalized, }{whether to normalized HHG to scale [0, 1], with TRUE as default}
}
\value{
correlation coefficients
}
\description{
A wrapper for HHG::HHG()
}
\details{
See help file for HHG::HHG

Official package HHG can be installed from CRAN or 
http://www.math.tau.ac.il/~ruheller/Software.html

Same code lines can also be found in the following site: 
https://github.com/soumikp/fastMI/blob/master/code/functions.R.gz
}
