% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_hoeffd.R
\name{cor_hoeffd}
\alias{cor_hoeffd}
\title{Hoeffding's D Statistics}
\usage{
cor_hoeffd(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for Hmisc::hoeffd()
}
\details{
See help file for Hmisc::hoeffd
}
