% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_hsic.R
\name{cor_hsic}
\alias{cor_hsic}
\title{HSIC}
\usage{
cor_hsic(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
HSIC
}
\details{
Attention: colculation on the same x and y, 
the resulted correlation coefficient will be different 
because of the random effect of kernlab::sigest


Codes from  
https://github.com/stochasticresearch/copulastatistic

which is further extracted from 

Code from Lopez-Paz 
https://github.com/lopezpaz/randomized_dependence_coefficient
}
