% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_hsiccop.R
\name{cor_hsiccop}
\alias{cor_hsiccop}
\title{CHSIC}
\usage{
cor_hsiccop(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
codes from github, see details
}
\details{
Codes from  
https://github.com/stochasticresearch/copulastatistic

which is further extracted from 

Code from Lopez-Paz 
https://github.com/lopezpaz/randomized_dependence_coefficient
}
