% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mcd.R
\name{cor_mcd}
\alias{cor_mcd}
\title{Minimum Covariance Determinant}
\usage{
cor_mcd(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for robustbase::covMcd()
}
\details{
See help file for robustbase::covMcd

There is cov.mcd() function in MASS,
which is slower
}
