% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mgc.R
\name{cor_mgc}
\alias{cor_mgc}
\title{Multiscale Graph Correlation}
\usage{
cor_mgc(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
sample distance correlation
}
\description{
A wrapper for mgc::mgc.stat
}
\details{
See help file for mgc::mgc.stat
}
\references{
J. T. Vogelstein, E. W. Bridgeford, Q. Wang, C. E. Priebe, M. Maggioni, C. Shen, 
Discovering and deciphering relationships across disparate data modalities. eLife. 8, e41690 (2019)
}
