% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mi.R
\name{cor_mi}
\alias{cor_mi}
\title{Mutual Information}
\usage{
cor_mi(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
Mutual Information Dependence Measure
}
\details{
There are several packages for MI, we adopt JMI here.

See JMI::JMI for more details.

\eqn{sqrt(1-exp(-2*mi))} is returned.
}
