% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mic.R
\name{cor_mic}
\alias{cor_mic}
\title{Maximal information coefficient}
\usage{
cor_mic(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for minerva::mine_stat(x, y, measure = 'mic')
}
\details{
See the help file for minerva::mine_stat
}
