% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_pearson.R
\name{cor_pearson}
\alias{cor_pearson}
\title{Fast implementations of (robust) correlation estimators}
\usage{
cor_pearson(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
Estimate the correlation of two vectors via fast C++ implementations, 
with a focus on robust and nonparametric methods.
}
\details{
Pearson's r is not calculated as cor(x, y). Instead, we adpot a 
faster version in wdm::wdm()

Until now (2022-11-14), ccaPP provides the fastest implementation 
of Pearson, Spearman, Kendall, Quadrant than 
stats::cor, wdm::wdm() 
wdm also claims being faster than cor
}
