% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_percentage_bend.R
\name{cor_percentage_bend}
\alias{cor_percentage_bend}
\title{Percentage bend correlation}
\usage{
cor_percentage_bend(x, y, beta = 0.2)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for asbio::r.pb
}
\details{
See help file for asbio::r.pb

The results is the same as WRS2::pbcor, 
but asbio::r.pb is faster.
}
\references{
R. R. Wilcox, Introduction to Robust Estimation and Hypothesis Testing, 3rd edtion. 
Academic Press, 2012. pp.446
}
