% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_rdc.R
\name{cor_rdc}
\alias{cor_rdc}
\title{The randomized dependence coefficient}
\usage{
cor_rdc(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
Code from the paper:
The randomized dependence coefficient
}
\details{
For more details, see D. Lopez-Paz, P. Hennig, B. Schölkopf, The randomized dependence coefficient, 
in: Advances in Neural Information Processing Systems, 2013: pp. 1–9.
}
