% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_RV.R
\name{cor_RV}
\alias{cor_RV}
\title{RV coefficient}
\usage{
cor_RV(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper of FactoMineR::coeffRV
}
\details{
See help file for FactoMineR::coeffRV

cor_RV(x, y) is actually cor_pearson(x, y)^2 when x and y are one-dimension variable

Obviously, when we calculate RV coefficient by cor_pearson_squared(x, y) 
rather than coeffRV(x, y)$rv, it will be much faster.
}
