% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tcor.R
\name{cor_tcor}
\alias{cor_tcor}
\title{Correlation with tighter bounds}
\usage{
cor_tcor(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
Faster
}
\details{
cor_tcor is the rearrangement correlation.

t stands for tighter bound, which leads to sharp values
}
