% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tic.R
\name{cor_tic}
\alias{cor_tic}
\title{Maximal Information Coefficient}
\usage{
cor_tic(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for minerva::mine_stat(x, y, measure = 'tic')
}
\details{
See help file for minerva::mine_stat
}
