% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_win.R
\name{cor_win}
\alias{cor_win}
\title{Winsorized correlation}
\usage{
cor_win(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper for WRS2::wincor
}
\details{
See help file for WRS2::wincor
}
