% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_xicor.R
\name{cor_xicor}
\alias{cor_xicor}
\title{Compute the cross rank increment correlation coefficient xi}
\usage{
cor_xicor(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
code from XICOR::calculateXI()
}
\details{
See help file for XICOR::calculateXI


https://github.com/zhexiaolin/Limit-theorems-of-Chatterjee-s-rank-correlation
}
