% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_xicor_c.R
\name{cor_xicor_c}
\alias{cor_xicor_c}
\title{Estimate the conditional dependence coefficient (CODEC) xi_cor}
\usage{
cor_xicor_c(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
A wrapper of FOCI::codec
}
\details{
See help file for FOCI::codec

The conditional dependence coefficient (CODEC) 
is a measure of the amount of conditional dependence 
between a random variable Y and a random vector Z 
given a random vector X, based on an i.i.d. sample of (Y, Z, X). 
The coefficient is asymptotically guaranteed to be between 0 and 1.
}
\references{
M. Azadkia, S. Chatterjee, A simple measure of conditional dependence. 
Ann. Statist. 49 (2021), doi:10.1214/21-AOS2073
}
