% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_rearrangement.R
\name{cov_rearrangement}
\alias{cov_rearrangement}
\title{bound of covariance - Rearrangement Covariance of x and y}
\usage{
cov_rearrangement(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
a numeric vector of length one
}
\description{
Rearrangement Covariance of x and y
}
\details{
rearrangement covariance of x and y is defined as 
cov(sort(x, T), sort(y, T)) if cov(x, y) >= 0
cov(sort(x, T), sort(y, F)) if cov(x, y) < 0
}
