% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dB_to_R2.R
\name{dB_to_R2}
\alias{dB_to_R2}
\title{Transitions among SNR,SNRdB, r and R^2}
\usage{
dB_to_R2(dB)
}
\arguments{
\item{dB, }{Signal-to-Noise Ration in decibels}
}
\value{
corresponding R^2 score of SNRdB as a numeric vector
}
\description{
transition between SNRdB and R^2
}
\details{
R2 <- SNR / (1 + SNR)
SNR <- R2 / (1 - R2)

SNRdB = 10*log10(SNR)
SNR = 10^(SNRdB/10)
}
