% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_code_str.R
\name{eval_code_str}
\alias{eval_code_str}
\title{Wrapper for eval code strings}
\usage{
eval_code_str(code_str)
}
\arguments{
\item{code_str, }{string to be evaluated}
}
\value{
The results of code_str as code
}
\description{
Wrapper for eval code strings
}
\details{
A wrapper for eval code strings
envir is set as parent.frame()

For internal use in sim_power()
}
