% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_y_funs.R
\name{funs_simon_Tibshirani}
\alias{funs_simon_Tibshirani}
\title{classical function strings}
\usage{
to do
}
\value{
popular function names
}
\description{
to do
}
\details{
funs_simon_Tibshirani(): functions utilized by simon Tibshirani

funs_reshef(): functions utilized by reshef
}
\references{
D. N. Reshef, Y. A. Reshef, P. C. Sabeti, M. Mitzenmacher, 
An empirical study of the maximal and total information coefficients 
and leading measures of dependence. Ann. Appl. Stat. 12, 123–155 (2018).
available at: https://doi.org/10.1214/17-AOAS1093SUPPA

H. Jiang, Q. Wu, Robust dependence measure for detecting associations 
in large data set. Acta Mathematica Scientia. 38, 57–72 (2018).
}
