% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_digits.R
\name{get_digits}
\alias{get_digits}
\title{Extract the digital parts from a character vector}
\usage{
get_digits(x)
}
\arguments{
\item{x, }{a character vector}
}
\value{
a numeric vector
}
\description{
Extract the digital parts from a character vector
}
\details{
as.numeric(str_extract(x, "\\d+\\.?\\d{0,}"))
}
