% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noised_y.R
\name{get_noised_y}
\alias{get_noised_y}
\title{Generate noised y with certain SNR}
\usage{
get_noised_y(y_hat, SNR = 1, dB = NULL)
}
\arguments{
\item{y_hat, }{signal, y_hat = f(x)}

\item{SNR, }{Signal-to-Noise Ration}

\item{dB, }{Signal-to-Noise Ration in decibels}
}
\value{
noised y, i.e., f(x) + noise
}
\description{
generate noise for certain signal at certain SNR
}
\details{
noise signal = signal + noise

y = y_hat + noise
}
