% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power_df.R
\name{get_power_df}
\alias{get_power_df}
\title{Get the power_df}
\usage{
get_power_df(results, 0.95)
}
\arguments{
\item{results, }{a list including null_resluts and alternative_results}

\item{alpha, }{confidence level, with 0.95 as default}
}
\value{
power data.frame
}
\description{
Calculate the power data.frame on a given confidence level
}
\details{
results is usually resulted by power_analysis()
recommended alpha values are 0.90, 0.95, and 0.99
}
