% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_A_in_B.R
\name{index_of_A_in_B}
\alias{index_of_A_in_B}
\title{To get the index of A in B}
\usage{
index_of_A_in_B(A, B)
}
\arguments{
\item{A, }{a vector}

\item{B, }{a vector}
}
\value{
the index of A in B
}
\description{
To get the index of A in B
}
\details{
A wrapper for match(A, B)
}
