% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all.R
\name{load_all}
\alias{load_all}
\title{Load all objects in several .rda files into a whole list}
\usage{
load_all(files_to_be_loaded)
}
\arguments{
\item{files_to_be_loaded, }{several file paths}
}
\value{
a list, consisting all the loaded objects
}
\description{
Load all objects in several .rda files into a whole list
}
\details{
all the objects are loaded into a list, 
each as a component.

For example, if there are 3 objects in file1 and 4 objects in file2,
then the resulted list will contain all these 7 objects
}
