% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rbind.data.table.R
\name{load_rbind.data.table}
\alias{load_rbind.data.table}
\title{Load data.table objects in different .rda files and rbind them together}
\usage{
load_rbind.data.table(dir = NULL, ...)
}
\arguments{
\item{dir, }{destination directory for rda files}

\item{..., }{more parameters to be passed to list.files, such as pattern}
}
\value{
a long data.table (data.frame)
}
\description{
Load data.table objects in different .rda files and rbind them together
}
\details{
if more than one data.frame objects are stored in some .rda file, 
only the first one will be loaded, and the remaining ones are omitted
}
