% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rbind.R
\name{load_rbind}
\alias{load_rbind}
\title{Load data.frame objects in different .rda files and rbind them together}
\usage{
load_rbind(dir = NULL)
}
\arguments{
\item{dir, }{destination directory for rda files}
}
\value{
a long data.frame
}
\description{
Load data.frame objects in different .rda files and rbind them together
}
\details{
if more than one data.frame objects are stored in some .rda file, 
only the first one will be loaded, and the remaining ones are omitted
}
