% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rbind_powerdf.R
\name{load_rbind_powerdf}
\alias{load_rbind_powerdf}
\title{Load and rbind the powerdf rda files}
\usage{
load_rbind_powerdf(pdir = NULL)
}
\arguments{
\item{pdir, }{the path storing the powerdf rda files}
}
\value{
a list
}
\description{
load several powerdf rda files all at once and rbind the power_df into a list
}
\details{
load_all() is called, and a list including three parts, power_df.90, .95 and .99 will be returned
}
