% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_settings.R
\name{load_settings}
\alias{load_settings}
\title{Default setting}
\usage{
load_settings()
}
\description{
Load default setting
}
\details{
Some constants shared in different functions:
\itemize{
  \item .base_family
  \item .base_size
  \item .dB_0.25
  \item .dB_0.50
  \item .dB_0.75
  \item .dB_1.00
  \item .dBs
  \item .lower_dB
  \item .Random.seed
  \item .upper_dB
}
}
