% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looser_pearson.R
\name{looser_pearson}
\alias{looser_pearson}
\title{loose version of Pearson's r}
\usage{
looser_pearson(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
Coefficient of identity
}
\description{
loose version of Pearson's r
}
\details{
Scale covariance with an even looser bound than that 
by Cauchy-Schwarz Inequality
}
