% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosest_pearson.R
\name{loosest_pearson}
\alias{loosest_pearson}
\title{Even looser version of Pearson's r, identical without shifts}
\usage{
loosest_pearson(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
Coefficient of identity
}
\description{
loose version of Pearson's r
}
\details{
Scale covariance with an even looser bound than that 
by Cauchy-Schwarz Inequality
}
