% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_plot_lite.R
\name{power_plot_lite}
\alias{power_plot_lite}
\title{Raincloud plots for power analysis}
\usage{
power_plot_lite(
  power_results = NULL,
  power_results_path = NULL,
  lower_dB = .lower_dB,
  upper_dB = .upper_dB,
  cor_types = NULL,
  scenario_type = c("power_distinguish", "all", "accuracy", "accuracy_mon",
    "accuracy_notmon", "power", "power_mon", "power_notmon", "extreme_mon_pairs"),
  omit_indistinguished_scenarios = FALSE,
  cor_colors = NULL,
  print = TRUE,
  show_exprs = TRUE,
  n_col = 3,
  cor_type_expr = supported_cor_expr(),
  dir = "v",
  ...
)
}
\arguments{
\item{power_results, }{the result of cal_power}

\item{power_results_path, }{if power_results is null, latest data will be loaded automatically from this directory}

\item{lower_dB, }{the lower bound of selected dB range, with -5 as default}

\item{upper_dB, }{the upper bound of selected dB range, with 5 as default}

\item{cor_types, }{selected correlation statistics to be visualized}

\item{scenario_type, }{which type of scenarios to be plotted}

\item{omit_indistinguished_scenarios, }{whether indistinguished scenarios are omitted, default TRUE}

\item{cor_colors, }{colors to distinguished whether correlation is sharp, with .six_favor_colors[c(1, 3)] as default}

\item{print, }{whether to print the plot}

\item{show_exprs, }{to show the expressions instead of cor_type strings, with TRUE as default}

\item{n_col, }{number of pannel columns, with 3 as default}

\item{cor_type_expr, }{cor_types and their corresponding expressions}

\item{dir, }{direction, with "v" as default}
}
\value{
a ggplot object
}
\description{
Raincloud plots, for comparison among different correlation statistics
}
\details{
The workflow for complete power analysis is as follows: 
\enumerate{
  \item Call power_db() to generate the basic results
  \item Call power_results_frm_db() to generate the power results data.frame
  \item Call power_plot_lite() to draw the figures
}

See more details at https://arxiv.org/abs/2205.04571

Make sure either power_results or power_results_path should be set correctly.
}
