% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_results_frm_db.R
\name{power_results_frm_db}
\alias{power_results_frm_db}
\title{Accuracy analysis for different correlation statistics}
\usage{
power_results_frm_db(
  power_db_dir = "./power_db",
  cor_types = c("loose_pearson", "cor_pearson", "sharp_pearson", "cor_spearman",
    "cor_kendall", "cor_dhsic", "cor_dcor", "cor_mic", "cor_xicor"),
  ...
)
}
\arguments{
\item{power_db_dir, }{the path of power_db, including several power_simulation_cor_type_fun1~107_timestamp.rda files}

\item{cor_types, }{correlation types, with c('loose_pearson', 'cor_pearson', 'sharp_pearson', 'cor_spearman',  'cor_kendall', 'cor_dhsic', 'cor_dcor', 'cor_mic', 'cor_xicor') as default}
}
\value{
power_results
}
\description{
Calculate the correlation coefficients for different types of functions
under different levels of noises 
to evaluate whether the correlation values are similar under certain noise
}
\details{
The workflow for complete power analysis is as follows: 
\enumerate{
  \item Call power_db() to generate the basic results
  \item Call power_results_frm_db() to generate the power results data.frame
  \item Call power_plot_lite() to draw the figures
}

See more details at https://arxiv.org/abs/2205.04571
}
