% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_to_dB.R
\name{R2_to_SNR}
\alias{R2_to_SNR}
\title{Transitions among SNR,SNRdB, r and R^2}
\usage{
R2_to_SNR(R2)
}
\arguments{
\item{R2, }{Coefficient of determination R^2}
}
\value{
corresponding SNRdB of the given R^2 as a numeric vector
}
\description{
transition between R2 and SNRdB#'
}
\details{
R2 <- SNR / (1 + SNR)
SNR <- R2 / (1 - R2)

SNRdB = 10*log10(SNR)
SNR = 10^(SNRdB/10)
}
