% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_arithmetic.R
\name{r_arithmetic}
\alias{r_arithmetic}
\alias{r_prime}
\alias{r_minus}
\title{The degenerated r}
\usage{
r_arithmetic(x, y)

r_prime(x, y)

r_minus(x, y)
}
\arguments{
\item{x, }{variable x}

\item{y, }{variable y}
}
\value{
correlation coefficient
}
\description{
The degenerated r
}
\details{
See https://arxiv.org/abs/2205.04571 for more details.
}
