% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_score_theoretical.R
\name{r_score_theoretical}
\alias{r_score_theoretical}
\title{Get the theoretical r scores with resepct to different SNRdB}
\usage{
r_score_theoretical(dB = -50:50)
}
\arguments{
\item{dB, }{Signal-to-Noise Ration in decibels, with -50:50 as default}
}
\value{
theoretical r scores in a numeric vector
}
\description{
get the theoretical r scores with resepct to different SNRdB
}
\details{
r = sqrt(1 / (1 + 10^(-0.1dB)))
}
