% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_to_dB.R
\name{R_to_dB}
\alias{R_to_dB}
\title{Conversions among dB, SNR, R2, R}
\usage{
R_to_dB(R)
}
\arguments{
\item{R, }{R}
}
\value{
dB
}
\description{
Conversions among dB, SNR, R2, R
}
\details{
Details are described in the manuscript, 
which is available at https://arxiv.org/abs/2205.04571
}
