% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_to_SNR.R
\name{R_to_SNR}
\alias{R_to_SNR}
\title{Conversions among dB, SNR, R2, R}
\usage{
R_to_SNR(R)
}
\arguments{
\item{R, }{R}
}
\value{
SNR
}
\description{
Conversions among dB, SNR, R2, R
}
\details{
See https://arxiv.org/abs/2205.04571 for details
}
