% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recor.R
\name{recor}
\alias{recor}
\title{REarrangement-type CORrelation - the adjusted r}
\usage{
recor(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
numeric vector of length one
}
\description{
correlation based on rearrangement covariance - the adjusted r
}
\details{
the adjusted r is defined as 
cov(x, y) / abs(cov_rearrangement(x, y))
}
