% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recor_pearson.R
\name{recor_pearson}
\alias{recor_pearson}
\title{Sharp version of cor_pearson}
\usage{
recor_pearson(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
Sharp version of cor_pearson
}
\details{
the same as cor_tcor
}
