% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_science.R
\name{science_colours_six}
\alias{science_colours_six}
\title{Colors}
\usage{
science_colours_six()
}
\value{
colors
}
\description{
Colors
}
\details{
Some colors used in paper: 
https://arxiv.org/abs/2205.04571
}
