% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharp_bcor.R
\name{sharp_bcor}
\alias{sharp_bcor}
\title{Sharp version of Ball Correlation - BCor}
\usage{
sharp_bcor(x, y)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
a numeric vector of length one
}
\description{
sharp version of Ball::bcor
}
\details{
The sharp BCor is defined as 
 BCov(x, y) / abs(BCov_rearrangement(x, y)). 
 Here, the BCov_rearrangement(x, y) is similar to cov_rearranement(x, y)
}
