% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharp_dcor.R
\name{sharp_dcor}
\alias{sharp_dcor}
\title{Sharp version of distance correlation - dCor}
\usage{
sharp_dcor(x, y, index = 1)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
a numeric vector of length one
}
\description{
sharp version of energy::dcor
}
\details{
The sharp dCor is defined as 
 dCov(x, y) / abs(dCov_rearrangement(x, y)). 
 Here, the dCov_rearrangement(x, y) is similar to cov_rearranement(x, y)
}
