% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power.R
\name{sim_power}
\alias{sim_power}
\title{Power Simulation in parallel}
\usage{
sim_power(
  x_strs = "runif(512)",
  y_strs = c("y <- x", "y <- x^3"),
  cor_strs = c("cor", "recor"),
  SNRs = dB_to_SNR(-50:50),
  nsim = 500,
  nround = 1,
  ncores = NULL,
  ...
)
}
\arguments{
\item{x_strs, }{the way to generate the independent variable x, with runif(512) as default}

\item{y_strs, }{y_strs, monotonic or non monotonic scenarios}

\item{cor_strs, }{cor_fun_strs, correlation statistics}

\item{SNRs, }{signal-to-noise ratio, with dB_to_SNR(-50:50) as default}

\item{nsim, }{sample length, with 500 as default}

\item{nround, }{number of simulation rounds, with 1 as default}

\item{ncores, }{cluster specification}
}
\value{
simulation results in a tidy data.frame
}
\description{
Power Simulation
}
\details{
x_strs: different sample size can be customized such as 
x_strs = paste0("runif", 10*50:100)

cor_strs: to calculate statistics such as dCor and MIC, energe and 
minerva should be loaded first
}
