% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNR_to_R.R
\name{SNR_to_R}
\alias{SNR_to_R}
\title{Conversions among dB, SNR, R2, R}
\usage{
SNR_to_R(SNR)
}
\arguments{
\item{SNR, }{SNR}
}
\value{
R
}
\description{
Conversions among dB, SNR, R2, R
}
\details{
Details are described in the manuscript, 
which is available at https://arxiv.org/abs/2205.04571
}
