% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_cor_expr.R
\name{supported_cor_expr}
\alias{supported_cor_expr}
\title{Supported cor types and their corresponding expressions}
\usage{
supported_cor_expr()
}
\value{
data.frame, one column as cor_type, the other as cor_expr
}
\description{
supported cor types and their corresponding expressions
}
