% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeit.R
\name{timeit}
\alias{timeit}
\title{To record the time}
\usage{
timeit(status = "continue", new_line = TRUE)
}
\arguments{
\item{status, }{a string vector of length 1 to label the task type}

\item{new_line, }{whether to add a new line on the console.}
}
\value{
Nothing. Only show the time elapsed in the console.
}
\description{
To record the time, often for logging purposes
}
\details{
status can be 
\enumerate{
  \item new, to start a new task and reclock
  \item continue, to continue an existing task
  \item off, to end the task and turn off the timer
}
}
