% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_cor_funs.R
\name{two_cor_funs}
\alias{two_cor_funs}
\title{Mostly used correlation functions}
\usage{
two_cor_funs()
}
\value{
2-7 correlaiton names
}
\description{
Mostly used correlation functions
}
\details{
two_cor_funs(): 'cor_spearman' and 'cor_tcor'  

three_cor_funs(): 'cor_pearson', 'cor_spearman', and 'cor_tcor'  

three_cor_funs2(): 'r_prime', 'r_pearson', and 'r_sharp'  

four_cor_funs(): 'cor_pearson', 'cor_spearman', 'cor_kendall', and 'cor_tcor'  

six_cor_funs(): cor_pearson,cor_spearman,cor_kendall,cor_tcor,cor_dcor, cor_mic  

six_cor_funs2(): r_prime, cor_pearson, cor_spearman, cor_tcor, cor_dcor, cor_mic  

seven_cor_funs()：union(six_cor_funs(), six_cor_funs2())
}
