% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x1_minus_x2.R
\name{x1_minus_x2}
\alias{x1_minus_x2}
\title{X1 - X2 for general correlation statistics}
\usage{
x1_minus_x2(x)
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
numeric vector
}
\description{
X1 - X2 for general correlation statistics
}
\details{
X1 - X2 is the same as 
that in a general correlation statistics 
proposed by Kendall#'
}
