% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dB_to_R2.R
\name{dB_to_R2}
\alias{dB_to_R2}
\title{Conversions among dB, SNR, R2, R}
\usage{
dB_to_R2(dB)
}
\arguments{
\item{dB, }{dB}
}
\value{
R2
}
\description{
Conversions among dB, SNR, R2, R
}
\details{
Details are described in the manuscript, 
which is available at https://arxiv.org/abs/2205.04571
}
