% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dB_to_SNR.R
\name{dB_to_SNR}
\alias{dB_to_SNR}
\title{Transitions among SNR,SNRdB, r and R^2}
\usage{
dB_to_SNR(dB)
}
\arguments{
\item{dB, }{Signal-to-Noise Ration in decibels}
}
\value{
corresponding SNR of SNRdB
}
\description{
transition between SNRdB and SNR
}
\details{
dB = 10*log10(SNR)
SNR = 10^(db/10)
}
