% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_to_dB.R
\name{R2_to_dB}
\alias{R2_to_dB}
\title{Conversions among dB, SNR, R2, R}
\usage{
R2_to_dB(R2)
}
\arguments{
\item{R2, }{R2}
}
\value{
dB
}
\description{
Conversions among dB, SNR, R2, R
}
\details{
See https://arxiv.org/abs/2205.04571 for details
}
