% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_to_SNR.R
\name{R2_to_SNR}
\alias{R2_to_SNR}
\title{Conversions among dB, SNR, R2, R}
\usage{
R2_to_SNR(R2)
}
\arguments{
\item{R2, }{R2}
}
\value{
SNR
}
\description{
Conversions among dB, SNR, R2, R
}
\details{
Details are described in the manuscript, 
which is available at https://arxiv.org/abs/2205.04571
}
