% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_score_theoretical.R
\name{r_score_theoretical}
\alias{r_score_theoretical}
\title{To get the theoretical r scores regarding to different SNR / dB}
\usage{
r_score_theoretical(dB = -50:50)
}
\arguments{
\item{dB, }{signal-to-noise ratio in decibels, with -50:50 as default}
}
\value{
power, scalar value
}
\description{
To get the theoretical r scores regarding to different SNR / dB
}
\details{
r = sqrt(1 / (1 + 10^(-0.1dB)))
}
