% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recor.R
\name{recor}
\alias{recor}
\title{Rearrangement correlation}
\usage{
recor(x, y, method = "cor_pearson")
}
\arguments{
\item{x, }{numeric vector}

\item{y, }{numeric vector}
}
\value{
correlation coefficients
}
\description{
to adjust the underestimation for monotonic dependence
}
\details{
This function is able to sharpen arbitrary correlation 
which is sensitive to monotonicity,
providing a simple and unified interface.

However, it is at the cost of efficiency.

For example, recor dcor will takes twice the time 
as dcov2d(x, y) / dcov2d(x_up, y_updown).
}
