% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power.R
\name{sim_power}
\alias{sim_power}
\title{Basic function for power analysis}
\usage{
sim_power(
  x_strs = "runif(512)",
  y_strs = c("y <- x", "y <- x^3"),
  cor_strs = three_cor_funs(),
  SNRs = dB_to_SNR(-50:50),
  nsim = 500,
  nround = 1,
  err_log = "err_log.rda",
  ...
)
}
\arguments{
\item{x_strs, }{the way to generate the independent variable x, with runif(512) as default}

\item{y_strs, }{y_strs such as y <- x}

\item{cor_strs, }{cor_fun_strs, with the returns of three_cor_funs() as default}

\item{SNRs, }{signal-to-noise ratio, with dB_to_SNR(-50:50) as default}

\item{nsim, }{times of null simulation, with 500 as default}
}
\value{
simulation results in a data.frame
}
\description{
Basic function for power analysis
}
\details{
The core function for power analysis, 
usually as a workhorse called by power_db()

For efficiency, parallel computation is adopted here.

The algorithm automatically detects how many logical cores can be employed,
and parallel::detectCores() - 1 cores are actually used for this task.
}
