% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNR_to_dB.R
\name{SNR_to_dB}
\alias{SNR_to_dB}
\title{Conversions among dB, SNR, R2, R}
\usage{
SNR_to_dB(SNR)
}
\arguments{
\item{SNR, }{SNR}
}
\value{
dB
}
\description{
Conversions among dB, SNR, R2, R
}
\details{
Details are described in the manuscript, 
which is available at https://arxiv.org/abs/2205.04571
}
