% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNR_to_dB.R
\name{SNR_to_dB}
\alias{SNR_to_dB}
\title{Transitions among SNR,SNRdB, r and R^2}
\usage{
SNR_to_dB(SNR)
}
\arguments{
\item{SNR, }{signal-to-noise ratio, var(y_hat + noise) / var(noise)}
}
\value{
corresponding SNRdB of SNR as a numeric vector
}
\description{
transition between SNR and SNRdB
}
\details{
R2 <- SNR / (1 + SNR)
SNR <- R2 / (1 - R2)

SNRdB = 10*log10(SNR)
SNR = 10^(SNRdB/10)
}
