% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNR_to_R2.R
\name{SNR_to_R2}
\alias{SNR_to_R2}
\title{Transitions among SNR,SNRdB, r and R^2}
\usage{
SNR_to_R2(SNR)
}
\arguments{
\item{SNR, }{signal-to-noise ratio, a numeric vector}
}
\value{
a vector of R2 values
}
\description{
transition between SNR and R^2
}
\details{
R2 <- SNR / (1 + SNR)
SNR <- R2 / (1 - R2)
}
