% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_cor_expr.R
\name{supported_cor_expr}
\alias{supported_cor_expr}
\title{supported cor types and their corresponding expressions}
\usage{
supported_cor_expr()
}
\value{
data.frame, one column as cor_type, the other as cor_expr
}
\description{
supported cor types and their corresponding expressions
}
\details{
all supported cor types and their corresponding expressions

including those sharp ones

\tabular{ll}{
  \strong{cor_type} \tab \strong{cor_expr} \cr
  cor_ACE               \tab ACE                                  \cr
  cor_Ahat              \tab widehat(italic(A))                   \cr
  cor_bcor              \tab BCor                                 \cr
  sharp_bcor            \tab BCor^"#"                             \cr
  cor_bicor             \tab BiCor                                \cr
  sharp_bicor           \tab BiCor^"#"                            \cr
  cor_blomqvist         \tab Blomqvist*"'"*s~beta                 \cr
  cor_Ccor              \tab Ccor                                 \cr
  cor_CoS               \tab CoS                                  \cr
  cor_copent            \tab CopulaEntropy                        \cr
  cor_dcor              \tab dCor                                 \cr
  sharp_dcor            \tab dCor^"#"                             \cr
  cor_dcor_squared      \tab dCor^2                               \cr
  cor_dcor_rank         \tab dCor[italic(rank)]                   \cr
  cor_dhsic             \tab italic(d)*HSIC                       \cr
  sharp_dhsic           \tab italic(d)*HSIC^"#"                   \cr
  cor_gaussrank         \tab GaussRank                            \cr
  cor_gmic              \tab GMIC                                 \cr
  cor_Hellinger         \tab HellingerCor                         \cr
  cor_HHG               \tab HHG                                  \cr
  sharp_HHG             \tab HHG^"#"                              \cr
  cor_HHG_norm          \tab HHG_norm                             \cr
  cor_hoeffd            \tab Hoeffding*"'"*s~italic(D)            \cr
  cor_hsic              \tab HSIC                                 \cr
  cor_hsiccop           \tab italic(c)*HSIC                       \cr
  cor_identical         \tab italic(r)^symbol("-")                \cr
  cor_kendall           \tab Kendall*"'"*s~italic(tau)            \cr
  cor_mcd               \tab MCD                                  \cr
  cor_mgc               \tab MGC                                  \cr
  cor_mi                \tab MI                                   \cr
  cor_mic               \tab MIC                                  \cr
  cor_mic_sqrt          \tab italic(MIC)^frac(1, 2)               \cr
  cor_pearson           \tab Pearson*"'"*s~italic(r)              \cr
  sharp_pearson         \tab italic(r)^symbol("#")                \cr
  cor_pearson_squared   \tab italic(r)^2                          \cr
  cor_percentage_bend   \tab PbCor                                \cr
  sharp_percentage_bend \tab italic(r)[pb]^symbol("#")            \cr
  cor_quadrant          \tab QuadrantCor                          \cr
  cor_rdc               \tab RDC                                  \cr
  cor_RV                \tab RV                                   \cr
  sharp_RV              \tab RV^symbol("#")                       \cr
  cor_spearman          \tab Spearman*"'"*s~italic(rho)           \cr
  cor_TauStar           \tab Bergsma*"'"*s~italic(tau)^symbol("*")\cr
  sharp_TauStar         \tab italic(tau)^symbol("*"~"#")          \cr
  cor_tcor              \tab italic(r)^symbol("#")                \cr
  cor_tic               \tab TIC                                  \cr
  cor_win               \tab WinCor                               \cr
  sharp_win             \tab italic(r)[italic(w)]^symbol("#")     \cr
  cor_xicor             \tab Chatterjee*"'"*s~italic(xi)          \cr
  sharp_xicor           \tab italic(xi)^symbol("#")               \cr
  cor_xicor_c           \tab CODEC                                \cr
  r_prime               \tab italic(r)^symbol("-")                \cr
  loose_r               \tab italic(r)^symbol("-")                \cr
  looser_r              \tab italic(r)^symbol("-")                \cr
  zoom_pearson025       \tab zoom_pearson025                      
}
}
